// src/app/model/common/Allegato.ts
export class Allegato {
  idDomandaIscrizione: number;
  codCondizione: string;
  idAnagraficaSoggetto: number;
  cognome: string;
  nome: string;
  codiceFiscaleSoggetto: string;
  codTipoSoggetto: string;
  idAllegato: number;
  nomeFile: string;
  mimeType: string;
  protocollo: string;
  codTipoAllegato: string;
  codTipoAllegatoRed: string;
  codiceFiscaleRichiedente: string;
  dataInserimento: string;
}